requirements_comfyui.txt
# ComfyUI Requirements for Hugging Face Spaces
torch>=2.0.0
torchvision>=0.15.0
torchaudio>=2.0.0
diffusers>=0.21.0
transformers>=4.30.0
accelerate>=0.20.0
safetensors>=0.3.0
pillow>=9.0.0
numpy>=1.24.0
requests>=2.28.0
websocket-client>=1.6.0
gradio>=4.0.0
huggingface_hub>=0.16.0
omegaconf>=2.3.0
timm>=0.9.0
opencv-python>=4.8.0
scipy>=1.10.0
scikit-image>=0.21.0
controlnet-aux>=0.0.6
xformers>=0.0.20
